+++ src/Elements/Column.jsx
@@ -1,0 +1,37 @@
+import { useDrop } from "react-dnd";
+
+const Column = ({ children, className, title }) => {
+  const [{ isOver, canDrop }, drop] = useDrop({
+    accept: "Our first type",
+    drop: () => ({ name: title }),
+    collect: (monitor) => ({
+      isOver: monitor.isOver(),
+      canDrop: monitor.canDrop(),
+    }),
+  });
+
+  const getBackgroundColor = () => {
+    if (isOver) {
+      if (canDrop) {
+        return "rgb(188,251,255)";
+      } else if (!canDrop) {
+        return "rgb(255,188,188)";
+      }
+    } else {
+      return "";
+    }
+  };
+
+  return (
+    <div
+      ref={drop}
+      className={className}
+      style={{ backgroundColor: getBackgroundColor() }}
+    >
+      <p>{title}</p>
+      {children}
+    </div>
+  );
+};
+
+export default Column
