+++ src/Pages/DragNDrop.jsx
@@ -5,144 +5,10 @@ import { DndProvider, useDrag, useDrop } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { COLUMN_NAMES } from "./constants";
 import { tasks } from "./tasks";
-import "./DragNDrop.css";
+import "./../css/DragNDrop.css";
+import MovableItem from "../Elements/MoveableItem";
+import Column from "../Elements/Column";
 
-const MovableItem = ({
-  name,
-  index,
-  currentColumnName,
-  moveCardHandler,
-  setItems,
-}) => {
-  const changeItemColumn = (currentItem, columnName) => {
-    setItems((prevState) => {
-      return prevState.map((e) => {
-        return {
-          ...e,
-          column: e.name === currentItem.name ? columnName : e.column,
-        };
-      });
-    });
-  };
-
-  const ref = useRef(null);
-
-  const [, drop] = useDrop({
-    accept: "Our first type",
-    hover(item, monitor) {
-      if (!ref.current) {
-        return;
-      }
-      const dragIndex = item.index;
-      const hoverIndex = index;
-      // Don't replace items with themselves
-      if (dragIndex === hoverIndex) {
-        return;
-      }
-      // Determine rectangle on screen
-      const hoverBoundingRect = ref.current?.getBoundingClientRect();
-      // Get vertical middle
-      const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-      // Determine mouse position
-      const clientOffset = monitor.getClientOffset();
-      // Get pixels to the top
-      const hoverClientY = clientOffset.y - hoverBoundingRect.top;
-      // Only perform the move when the mouse has crossed half of the items height
-      // When dragging downwards, only move when the cursor is below 50%
-      // When dragging upwards, only move when the cursor is above 50%
-      // Dragging downwards
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        return;
-      }
-      // Dragging upwards
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        return;
-      }
-      // Time to actually perform the action
-      moveCardHandler(dragIndex, hoverIndex);
-      // Note: we're mutating the monitor item here!
-      // Generally it's better to avoid mutations,
-      // but it's good here for the sake of performance
-      // to avoid expensive index searches.
-      item.index = hoverIndex;
-    },
-  });
-
-  const [{ isDragging }, drag] = useDrag({
-    item: { index, name, currentColumnName, type: "Our first type" },
-    end: (item, monitor) => {
-      const dropResult = monitor.getDropResult();
-
-      if (dropResult) {
-        const { name } = dropResult;
-        const { DO_IT, IN_PROGRESS, AWAITING_REVIEW, DONE } = COLUMN_NAMES;
-        switch (name) {
-          case IN_PROGRESS:
-            changeItemColumn(item, IN_PROGRESS);
-            break;
-          case AWAITING_REVIEW:
-            changeItemColumn(item, AWAITING_REVIEW);
-            break;
-          case DONE:
-            changeItemColumn(item, DONE);
-            break;
-          case DO_IT:
-            changeItemColumn(item, DO_IT);
-            break;
-          default:
-            break;
-        }
-      }
-    },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging(),
-    }),
-  });
-
-  const opacity = isDragging ? 0.4 : 1;
-
-  drag(drop(ref));
-
-  return (
-    <div ref={ref} className="movable-item" style={{ opacity }}>
-      {name}
-    </div>
-  );
-};
-
-const Column = ({ children, className, title }) => {
-  const [{ isOver, canDrop }, drop] = useDrop({
-    accept: "Our first type",
-    drop: () => ({ name: title }),
-    collect: (monitor) => ({
-      isOver: monitor.isOver(),
-      canDrop: monitor.canDrop(),
-    }),
-  });
-
-  const getBackgroundColor = () => {
-    if (isOver) {
-      if (canDrop) {
-        return "rgb(188,251,255)";
-      } else if (!canDrop) {
-        return "rgb(255,188,188)";
-      }
-    } else {
-      return "";
-    }
-  };
-
-  return (
-    <div
-      ref={drop}
-      className={className}
-      style={{ backgroundColor: getBackgroundColor() }}
-    >
-      <p>{title}</p>
-      {children}
-    </div>
-  );
-};
 
 export default function DragNDrop() {
   const [items, setItems] = useState(tasks);
